Made some changes to AddContactTest and BlockContactInteractorTest while also renaming the latter to BlockContactTest
